fix(refactor): commit of leftovers
