+++ honestcash-v2/src/app/modules/editor/components/editor/editor.component.ts
@@ -5,12 +5,13 @@ import {AppStates, selectEditorState} from '../../../../app.states';
 import {forkJoin, Observable, Subscription} from 'rxjs';
 import {EDITOR_STATUS, State as EditorState} from '../../../../store/editor/editor.state';
 import {EditorLoad, EditorStoryPropertyChange, EditorUnload} from '../../../../store/editor/editor.actions';
-import {EditorService, STORY_PROPERTIES} from '../../services/editor.service';
+import {EditorService} from '../../services/editor.service';
 import {Block} from '../../converters/json-to-html';
 import {isPlatformBrowser} from '@angular/common';
 import {ScriptService} from 'ngx-script-loader';
 import {concatMap} from 'rxjs/operators';
 import {EDITOR_EDITING_MODES} from '../header/header.component';
+import {STORY_PROPERTIES} from '../../shared/editor.story-properties';
 
 // @todo editor-v2: get rid of the factory provider pattern, it should be immutable!
 export const EDITOR_AUTO_SAVE = {
